Simplify source for facts, convos on external integration
