+++ backend/utils/memories/process_memory.py
@@ -22,16 +22,16 @@ from models.task import Task, TaskStatus, TaskAction, TaskActionProvider
 from models.trend import Trend
 from models.notification_message import NotificationMessage
 from utils.apps import get_available_apps
-from utils.llm import obtain_emotional_message, retrieve_metadata_fields_from_transcript
-from utils.llm import summarize_open_glass, get_transcript_structure, generate_embedding, \
+from utils.llm import obtain_emotional_message, retrieve_metadata_fields_from_transcript, \
+    summarize_open_glass, get_transcript_structure, generate_embedding, \
     get_plugin_result, should_discard_memory, summarize_experience_text, new_facts_extractor, \
     trends_extractor, get_email_structure, get_post_structure, get_message_structure, \
-    retrieve_metadata_from_email, retrieve_metadata_from_post, retrieve_metadata_from_message, retrieve_metadata_from_text
+    retrieve_metadata_from_email, retrieve_metadata_from_post, retrieve_metadata_from_message, retrieve_metadata_from_text, \
+    extract_facts_from_text
 from utils.notifications import send_notification
 from utils.other.hume import get_hume, HumeJobCallbackModel, HumeJobModelPredictionResponseModel
 from utils.retrieval.rag import retrieve_rag_memory_context
 from utils.webhooks import memory_created_webhook
-from utils.memories.facts import extract_facts_from_text
 
 
 def _get_structured(
